requirements.txt
# flask
Flask==0.11.1
Flask-LogConfig==0.4.2
Flask-Testing==0.6.1

# celery
celery==4.0.2
boto==2.42.0
gevent==1.1.2
kombu==4.0.2
librabbitmq==1.6.1

# redis
redis==2.10.5

# utils
envoy==0.0.3
requests==2.2.1
jsonpickle==0.9.3
parse==1.6.6
psutil==5.0.1

# logs
pyfluent==0.2.0
python-logstash==0.4.6

# tests
PyHamcrest==1.9.0
nose==1.3.7